+++ etc/sequelize/models/privileges.js
@@ -7,7 +7,7 @@ const { DataTypes } = require('sequelize');
  */
 function define(sequelize) {
     return sequelize.define('privileges', {
-        username: { type: DataTypes.STRING, primaryKey: true },
+        userid: { type: DataTypes.INTEGER, primaryKey: true },
         user_mgmt_priv: { type: DataTypes.BOOLEAN, defaultValue: false },
         data_mgmt_priv: { type: DataTypes.BOOLEAN, defaultValue: false },
     });